refactor util logic into ns logic
