+++ controller/clustercontroller.go
@@ -21,8 +21,8 @@ import (
 	crv1 "github.com/crunchydata/postgres-operator/apis/cr/v1"
 	"github.com/crunchydata/postgres-operator/config"
 	"github.com/crunchydata/postgres-operator/kubeapi"
+	"github.com/crunchydata/postgres-operator/ns"
 	"github.com/crunchydata/postgres-operator/operator"
-	"github.com/crunchydata/postgres-operator/util"
 
 	clusteroperator "github.com/crunchydata/postgres-operator/operator/cluster"
 	log "github.com/sirupsen/logrus"
@@ -58,7 +58,7 @@ func (c *PgclusterController) Run() error {
 
 // watchPgclusters is the event loop for pgcluster resources
 func (c *PgclusterController) watchPgclusters(ctx context.Context) error {
-	nsList := util.GetNamespaces(c.PgclusterClientset, operator.InstallationName)
+	nsList := ns.GetNamespaces(c.PgclusterClientset, operator.InstallationName)
 
 	for i := 0; i < len(nsList); i++ {
 		log.Infof("starting pgcluster controller for ns [%s]", nsList[i])
